No functional change.
No really: this is functionally identical to master, despite adding 100 LOC! The commit splits the newDepth == 2 case off of the general LMR code. The form of the implementation has changed from "math" to "logic" but is functionally identical. I have validated that the result is identical using very long runs of deep searches (~20xLTC). This was necessary because "bench" does not hit some of the cases.

This version has simpler code that is probably identical to master in functionality, although it lops off some conditions. After testing, I am pretty sure that those conditions cannot be met.

The code is also simpler than the previous, although still not simpler than master.

However, it no longer benchmarks as faster. So the success of this venture will depend on tuning.

The change has 3 separate benefits, which I will work to capitalize over the coming weeks:

- It exposes the logic of newDepth=2, which can be optimized independently. (Expect faster, simpler and better.)
- It exposes the parameters of newDepth=2, which can be optimized using SPSA.
- The newDepth > 2 case can be optimized without needing to account for the shallowest depth.