No functional change.
No really: this is functionally identical to master, despite adding 100 LOC! The commit splits the newDepth == 2 case off of the general LMR code. The form of the implementation has changed from "math" to "logic" but is functionally identical. I have validated that the result is identical using very long runs of deep searches (~20xLTC). This was necessary because "bench" does not hit some of the cases.

The change has 4 separate benefits, which I will work to capitalize over the coming weeks:

- It is faster. About 0.5% (benchmark below). If my understanding is correct, it should be green at STC and LTC simply because it is 0.5% faster.
- It exposes the logic of newDepth=2, which can be optimized independently. (Expect faster, simpler and better.)
- It exposes the parameters of newDepth=2, which can be optimized using SPSA.
- The newDepth > 2 case can be optimized without needing to account for the shallowest depth.

Results for 40 tests for each version:

            Base      Test      Diff
    Mean    1995605   2004231   -8626
    StDev   10482     11272     7504

p-value: 0.875
speedup: 0.004