Do not reduce castling a a side-effect of avoiding a see_ge call in LMR. Bench: 5863907
Posted to https://groups.google.com/forum/#!topic/fishcooking/uyOR1MpzUPQ. I think this change fixes a small bug, and no one disagreed. (I.e., there were no comments at all.)

The purpose of type_of(move) == NORMAL is to filter castling moves from the pos.see_ge() call.

The result of this implementation is that castling moves are reduced without any special consideration. That is, strictly based on where they occur in the move list.

Castling is generated in the move picker's quiet stage, where moves are sorted by history data:

      m.value =  cmh[pos.moved_piece(m)][to_sq(m)]
               + fmh[pos.moved_piece(m)][to_sq(m)]
               + fm2[pos.moved_piece(m)][to_sq(m)]
               + history[c][from_to(m)];

The encoding for castling as "KxR" means that the to_sq(move) is the rook square, which is A1 or H1. That is, the sorting order for OO is the same as for Kh1 and for OOO is the same as Ka1.

Ka1 and Kh1 are very seldom refutations, so this could cause OO and OOO to have low scores. But the situation is not too bad. Kh1/Ka1 is not playable until after OO/OOO has been played. So there is a phase of the game where castling is possible and Ka1/OOO and Kh1/OO will be confused, but eventually good castling moves end up in the TT, and all of this LMR stuff becomes irrelevant.

So there is a small issue here: the search reduces castling even though it is very often the best play, and it is a little slow to catch on.

The change codes the exception as follows:

              else if (type_of(move) == CASTLING
                  || !pos.see_ge(make_move(to_sq(move), from_sq(move))))
                  r -= 2 * ONE_PLY;

Compared to the original code, this also prevents castling moves from entering pos.see_ge(). In this case, it does this by reducing castling moves by two ply less, as if the move escaped a capture.

Since we are in the case !captureOrPromotion, ENPASSANT and PROMOTION are already excluded. Therefore, the proposed code searches identically to the original after castling rights are gone for both sides.

While the sides have castling rights, the proposed code searches castling moves up to 2 ply deeper. Note that castling is still possible for both sides in the vast majority of our suite of opening positions.