Split out a King and Pawn evaluation function.
The rationale is to provide a basis for the K&P evaluation function to evolve independently. That is, the program can have K&P parameters that vary, or even entirely different features. But that is for the future. Current commit has same signature, to benchmark behavior before trying to improve anything.

Bench: 3618154